feat(dataset): rearrange data schema folder
