+++ backend/api/schema/data/spectrogram__query.py
@@ -1,8 +1,8 @@
 """Spectrogram schema"""
-from graphene import relay
+from graphene import relay, ObjectType
 
 from backend.api.models import Spectrogram
-from backend.utils.schema import ApiObjectType
+from backend.utils.schema import ApiObjectType, AuthenticatedDjangoConnectionField
 
 
 class SpectrogramNode(ApiObjectType):
@@ -14,3 +14,9 @@ class SpectrogramNode(ApiObjectType):
         fields = "__all__"
         filter_fields = "__all__"
         interfaces = (relay.Node,)
+
+
+class SpectrogramQuery(ObjectType):
+    """Spectrogram queries"""
+
+    all_spectrograms = AuthenticatedDjangoConnectionField(SpectrogramNode)
