+++ frontend/schema.graphql
@@ -1114,7 +1114,7 @@ type Query {
     """Ordering"""
     orderBy: String
   ): SpectrogramAnalysisNodeNodeConnection
-  allDatasets(
+  allLegacySpectrogramConfigurations(
     """Query limit"""
     limit: Int
 
@@ -1127,28 +1127,7 @@ type Query {
     after: String
     first: Int
     last: Int
-
-    """Ordering"""
-    orderBy: String
-  ): DatasetNodeNodeConnection
-  datasetById(id: ID!): DatasetNode
-  allDatasetsAvailableForImport: [ImportDatasetType]
-  _debug: DjangoDebug
-  allColormaps(
-    """Query limit"""
-    limit: Int
-
-    """Query offset"""
-    offset: Int
-
-    """Query order"""
-    ordering: String
-    before: String
-    after: String
-    first: Int
-    last: Int
-    name: String
-  ): ColormapNodeNodeConnection
+  ): LegacySpectrogramConfigurationNodeNodeConnection
   allFfts(
     """Query limit"""
     limit: Int
@@ -1170,7 +1149,7 @@ type Query {
     scaling: String
     legacy: Boolean
   ): FFTNodeNodeConnection
-  allLegacySpectrogramConfigurations(
+  allDatasets(
     """Query limit"""
     limit: Int
 
@@ -1183,7 +1162,27 @@ type Query {
     after: String
     first: Int
     last: Int
-  ): LegacySpectrogramConfigurationNodeNodeConnection
+
+    """Ordering"""
+    orderBy: String
+  ): DatasetNodeNodeConnection
+  datasetById(id: ID!): DatasetNode
+  allDatasetsAvailableForImport: [ImportDatasetType]
+  allColormaps(
+    """Query limit"""
+    limit: Int
+
+    """Query offset"""
+    offset: Int
+
+    """Query order"""
+    ordering: String
+    before: String
+    after: String
+    first: Int
+    last: Int
+    name: String
+  ): ColormapNodeNodeConnection
   allLinearScales(
     """Query limit"""
     limit: Int
@@ -1531,6 +1530,7 @@ type Query {
     description: String
     labels: ID
   ): LabelSetNodeNodeConnection
+  _debug: DjangoDebug
 }
 
 type LabelNodeNodeConnection {
@@ -3829,6 +3829,24 @@ type UserNodeNodeConnection {
   totalCount: Int
 }
 
+type LegacySpectrogramConfigurationNodeNodeConnection {
+  """Pagination data for this connection."""
+  pageInfo: PageInfoExtra!
+
+  """Contains the nodes in this connection."""
+  results: [LegacySpectrogramConfigurationNode]!
+  totalCount: Int
+}
+
+type FFTNodeNodeConnection {
+  """Pagination data for this connection."""
+  pageInfo: PageInfoExtra!
+
+  """Contains the nodes in this connection."""
+  results: [FFTNode]!
+  totalCount: Int
+}
+
 type DatasetNodeNodeConnection {
   """Pagination data for this connection."""
   pageInfo: PageInfoExtra!
@@ -3852,96 +3870,12 @@ type ImportSpectrogramAnalysisType {
   path: String!
 }
 
-"""Debugging information for the current query."""
-type DjangoDebug {
-  """Executed SQL queries for this API query."""
-  sql: [DjangoDebugSQL]
-
-  """Raise exceptions for this API query."""
-  exceptions: [DjangoDebugException]
-}
-
-"""Represents a single database query made to a Django managed DB."""
-type DjangoDebugSQL {
-  """The type of database being used (e.g. postrgesql, mysql, sqlite)."""
-  vendor: String!
-
-  """The Django database alias (e.g. 'default')."""
-  alias: String!
-
-  """The actual SQL sent to this database."""
-  sql: String
-
-  """Duration of this database query in seconds."""
-  duration: Float!
-
-  """The raw SQL of this query, without params."""
-  rawSql: String!
-
-  """JSON encoded database query parameters."""
-  params: String!
-
-  """Start time of this database query."""
-  startTime: Float!
-
-  """Stop time of this database query."""
-  stopTime: Float!
-
-  """Whether this database query took more than 10 seconds."""
-  isSlow: Boolean!
-
-  """Whether this database query was a SELECT."""
-  isSelect: Boolean!
-
-  """Postgres transaction ID if available."""
-  transId: String
-
-  """Postgres transaction status if available."""
-  transStatus: String
-
-  """Postgres isolation level if available."""
-  isoLevel: String
-
-  """Postgres connection encoding if available."""
-  encoding: String
-}
-
-"""Represents a single exception raised."""
-type DjangoDebugException {
-  """The class of the exception"""
-  excType: String!
-
-  """The message of the exception"""
-  message: String!
-
-  """The stack trace"""
-  stack: String!
-}
-
 type ColormapNodeNodeConnection {
   """Pagination data for this connection."""
   pageInfo: PageInfoExtra!
 
   """Contains the nodes in this connection."""
   results: [ColormapNode]!
-  totalCount: Int
-}
-
-type FFTNodeNodeConnection {
-  """Pagination data for this connection."""
-  pageInfo: PageInfoExtra!
-
-  """Contains the nodes in this connection."""
-  results: [FFTNode]!
-  totalCount: Int
-}
-
-type LegacySpectrogramConfigurationNodeNodeConnection {
-  """Pagination data for this connection."""
-  pageInfo: PageInfoExtra!
-
-  """Contains the nodes in this connection."""
-  results: [LegacySpectrogramConfigurationNode]!
   totalCount: Int
 }
 
@@ -4087,6 +4021,72 @@ type LabelSetNodeNodeConnection {
   """Contains the nodes in this connection."""
   results: [LabelSetNode]!
   totalCount: Int
+}
+
+"""Debugging information for the current query."""
+type DjangoDebug {
+  """Executed SQL queries for this API query."""
+  sql: [DjangoDebugSQL]
+
+  """Raise exceptions for this API query."""
+  exceptions: [DjangoDebugException]
+}
+
+"""Represents a single database query made to a Django managed DB."""
+type DjangoDebugSQL {
+  """The type of database being used (e.g. postrgesql, mysql, sqlite)."""
+  vendor: String!
+
+  """The Django database alias (e.g. 'default')."""
+  alias: String!
+
+  """The actual SQL sent to this database."""
+  sql: String
+
+  """Duration of this database query in seconds."""
+  duration: Float!
+
+  """The raw SQL of this query, without params."""
+  rawSql: String!
+
+  """JSON encoded database query parameters."""
+  params: String!
+
+  """Start time of this database query."""
+  startTime: Float!
+
+  """Stop time of this database query."""
+  stopTime: Float!
+
+  """Whether this database query took more than 10 seconds."""
+  isSlow: Boolean!
+
+  """Whether this database query was a SELECT."""
+  isSelect: Boolean!
+
+  """Postgres transaction ID if available."""
+  transId: String
+
+  """Postgres transaction status if available."""
+  transStatus: String
+
+  """Postgres isolation level if available."""
+  isoLevel: String
+
+  """Postgres connection encoding if available."""
+  encoding: String
+}
+
+"""Represents a single exception raised."""
+type DjangoDebugException {
+  """The class of the exception"""
+  excType: String!
+
+  """The message of the exception"""
+  message: String!
+
+  """The stack trace"""
+  stack: String!
 }
 
 """Global mutation"""
